refactor: rearrange, remove stuff
