+++ generator/accessor/src/main/kotlin/me/kcra/takenaka/generator/accessor/AccessorGenerator.kt
@@ -17,6 +17,9 @@
 
 package me.kcra.takenaka.generator.accessor
 
+import kotlinx.coroutines.CoroutineName
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import me.kcra.takenaka.core.Workspace
 import me.kcra.takenaka.core.mapping.MutableMappingsMap
 import me.kcra.takenaka.core.mapping.adapter.replaceCraftBukkitNMSVersion
@@ -60,13 +63,15 @@ class AccessorGenerator(override val workspace: Workspace, val config: AccessorC
 
         generationContext(config.flavor) {
             config.accessors.forEach { classAccessor ->
-                val node = tree[classAccessor.internalName]
-                if (node == null) {
-                    logger.warn { "did not find class ancestry node with name ${classAccessor.internalName}" }
-                    return@forEach
-                }
+                launch(Dispatchers.Default + CoroutineName("generate-coro")) {
+                    val node = tree[classAccessor.internalName]
+                    if (node == null) {
+                        logger.warn { "did not find class ancestry node with name ${classAccessor.internalName}" }
+                        return@launch
+                    }
 
-                generateClass(classAccessor, node)
+                    generateClass(classAccessor, node)
+                }
             }
         }
     }
