+++ generator/web/src/main/kotlin/me/kcra/takenaka/generator/web/SignatureFormatter.kt
@@ -586,40 +586,6 @@ class SignatureFormatter : SignatureVisitor {
 }
 
 /**
- * Formats a modifier integer into a string.
- *
- * @param mod the modifier integer
- * @param mask the modifier mask (you can get that from the [Modifier] class or use 0)
- * @return the modifier string
- */
-fun formatModifiers(mod: Int, mask: Int): String = buildString {
-    val mMod = mod and mask
-
-    if ((mMod and Opcodes.ACC_PUBLIC) != 0) append("public ")
-    if ((mMod and Opcodes.ACC_PRIVATE) != 0) append("private ")
-    if ((mMod and Opcodes.ACC_PROTECTED) != 0) append("protected ")
-    if ((mMod and Opcodes.ACC_STATIC) != 0) append("static ")
-    // enums can have an abstract modifier (methods included) if its constants have a custom impl
-    // TODO: should we remove that?
-
-    // an interface is implicitly abstract
-    // we need to check the unmasked modifiers here, since ACC_INTERFACE is not among Modifier#classModifiers
-    if ((mMod and Opcodes.ACC_ABSTRACT) != 0 && (mod and Opcodes.ACC_INTERFACE) == 0) append("abstract ")
-    if ((mMod and Opcodes.ACC_FINAL) != 0) {
-        // enums and records are implicitly final
-        // we need to check the unmasked modifiers here, since ACC_ENUM is not among Modifier#classModifiers
-        if (mask != Modifier.classModifiers() || ((mod and Opcodes.ACC_ENUM) == 0 && (mod and Opcodes.ACC_RECORD) == 0)) {
-            append("final ")
-        }
-    }
-    if ((mMod and Opcodes.ACC_NATIVE) != 0) append("native ")
-    if ((mMod and Opcodes.ACC_STRICT) != 0) append("strict ")
-    if ((mMod and Opcodes.ACC_SYNCHRONIZED) != 0) append("synchronized ")
-    if ((mMod and Opcodes.ACC_TRANSIENT) != 0) append("transient ")
-    if ((mMod and Opcodes.ACC_VOLATILE) != 0) append("volatile ")
-}
-
-/**
  * Remaps a class name and creates a link if a mapping has been found.
  *
  * @param internalName the internal name of the class to be remapped
